Moved components to their respective parts.
