+++ src/WeekendBot/Services/ConsoleLoggingService.cs
@@ -20,7 +20,6 @@ using System.Threading.Tasks;
 using Discord.Common;
 using Discord.Common.Options;
 using Microsoft.Extensions.Options;
-using WeekendBot.Core;
 using WeekendBot.Utils;
 
 namespace WeekendBot.Services;
