+++ test/WeekendBot.Components.Test/WeekendInquiryServiceTest.cs
@@ -18,6 +18,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Discord.Common;
 using Discord.Common.Options;
 using Microsoft.Extensions.Options;
 using NSubstitute;
